Experimented a little with using the Vector type
In the fashion that it was actually supposed to be used.

Looks like the overhead is slightly costly in on relatively small images, but the gains from SIMD seem to be significant on some of the larger images.  Either that,  or maybe just the IO costs were in a different place or something.