Finished creating first naive version
Probably could use some optimisation yet.

I added some benchmarking for the naive, and it appears to be upwards of 4 times slower than my F# version, which makes sense as the F# naive is parallelised across the 4 cores of my machine, whereas this one is purely sequential.

Also, moved things around a bit because that seemed to be expected.