Modified the naive file to use the bounded for loop approach
Don't know why I didn't think of it a LOT sooner

Will need to benchmark it yet.

Something to consider for the future:  Will something like Radix sort or counting sort be faster for the sorting of the windows than using .NET's built-in sort?  The windows are typically quite small, so it be the case that something simple like Insertion sort could actually be the fastest.  What can be usefully vectorised, though?

Also, another thing to consider is whether it is possible to change from allocating a new array in every processWindow call, to using a pre-allocated pool of arrays.  Might just make things run faster and/or require less memory.