Added in and adapted Array.choose for ValueOption
Had to modify it a little to take into account the fact that I don't have access to some of the library-internal functions

Even so, it seems to run faster.  A quick test on the medium size image with a window size of 5 shows a roughly 20% decrease in running time per image.