Modified the benchmarking program to test the divison alternatives
And tried to use the Disassembly options in BenchmarkDotNet.  Unfortunately, I don't think I was using them correctly, as the output doesn't really look like what they show on the website (and for that matter, it isn't clear to me what it is doing...)

In terms of the benchmarking results themselves, it is highly inconclusive.  Sometimes one method seems to come out a lot faster than the other, but in the next test the opposite is true.  Or, they're just about even.

Also, tidied up the Naive program a little.