Tidied up the program
Took some advice on Slack from Dave.Curlyo and others, and added a few type declarations, as well as compacted a lot of the stuff that keeps getting passed through down into one record type.

It has made the program longer in terms of LoC, but arguably much easier to read.