Change cml implementation
Restructured the whole thing.

Certainly nicer to look at, probably should be faster too.

Currently a bit of an issue in that in order to make the proxels, the proxels need to be made already...

Going to switch that to a two-stage process, so that the proxels are created, and then their alternatives updated to be correct.