Change CML to use a global Alt.never
Uses the Lazy instantiation process.  Basically, every instance of Alt.never in the arrays is actually the exact same one.

This appears to have led to a small improvment of roughly 5% on performance (not properly benchmarked, however).

This assessment was based only on the very small test image and 5 x 5 window, so results may be different with larger images or windows.